+++ dflib/src/main/java/org/dflib/IntSeries.java
@@ -1,5 +1,6 @@
 package org.dflib;
 
+import org.dflib.agg.Variance;
 import org.dflib.builder.BoolBuilder;
 import org.dflib.op.ReplaceOp;
 import org.dflib.series.FalseSeries;
@@ -389,33 +390,23 @@ public interface IntSeries extends Series<Integer> {
     }
 
     /**
-     * Compute the standard deviation, using the population variant
+     * Computes the standard deviation, using the population variant
      */
     default double stdDev() {
         return stdDev(true);
     }
 
     /**
-     * Compute the variance
+     * Computes the variance
      *
      * @param usePopulationVariance Use the Population variant if true, Sample variant if false
      */
     default double variance(boolean usePopulationVariance) {
-        int len = size();
-        double avg = avg();
-        double denominator = usePopulationVariance ? len : len - 1;
-
-        double acc = 0;
-        for (int i = 0; i < len; i++) {
-            final double x = this.getInt(i);
-            acc += (x - avg) * (x - avg);
-        }
-
-        return acc / denominator;
+        return Variance.ofIntSeries(this, usePopulationVariance);
     }
 
     /**
-     * Compute the variance, using the population variant
+     * Computes the variance, using the population variant
      */
     default double variance() {
         return variance(true);