+++ dflib/src/main/java/org/dflib/agg/Variance.java
@@ -1,5 +1,9 @@
 package org.dflib.agg;
 
+import org.dflib.DoubleSeries;
+import org.dflib.FloatSeries;
+import org.dflib.IntSeries;
+import org.dflib.LongSeries;
 import org.dflib.Series;
 
 import java.math.BigDecimal;
@@ -11,6 +15,88 @@ import java.math.MathContext;
  */
 public class Variance {
 
+    // TODO: replace ofXSeries with array access that can be invoked from specific series (e.g. see Percentiles.ofArray(..))
+
+    public static double ofIntSeries(IntSeries s, boolean usePopulationVariance) {
+        int len = s.size();
+        if (len == 0) {
+            return Double.NaN;
+        } else if (len == 1) {
+            return usePopulationVariance ? 0. : Double.NaN;
+        }
+
+        double avg = s.avg();
+        double denominator = usePopulationVariance ? len : len - 1;
+
+        double acc = 0;
+        for (int i = 0; i < len; i++) {
+            final double x = s.getInt(i);
+            acc += (x - avg) * (x - avg);
+        }
+
+        return acc / denominator;
+    }
+
+    public static double ofLongSeries(LongSeries s, boolean usePopulationVariance) {
+        int len = s.size();
+        if (len == 0) {
+            return Double.NaN;
+        } else if (len == 1) {
+            return usePopulationVariance ? 0. : Double.NaN;
+        }
+
+        double avg = s.avg();
+        double denominator = usePopulationVariance ? len : len - 1;
+
+        double acc = 0;
+        for (int i = 0; i < len; i++) {
+            final double x = s.getLong(i);
+            acc += (x - avg) * (x - avg);
+        }
+
+        return acc / denominator;
+    }
+
+    public static double ofFloatSeries(FloatSeries s, boolean usePopulationVariance) {
+        int len = s.size();
+        if (len == 0) {
+            return Double.NaN;
+        } else if (len == 1) {
+            return usePopulationVariance ? 0. : Double.NaN;
+        }
+
+        float avg = s.avg();
+        double denominator = usePopulationVariance ? len : len - 1;
+
+        double acc = 0;
+        for (int i = 0; i < len; i++) {
+            float x = s.getFloat(i);
+            acc += (x - avg) * (x - avg);
+        }
+
+        return acc / denominator;
+    }
+
+    public static double ofDoubleSeries(DoubleSeries s, boolean usePopulationVariance) {
+        int len = s.size();
+        if (len == 0) {
+            return Double.NaN;
+        } else if (len == 1) {
+            return usePopulationVariance ? 0. : Double.NaN;
+        }
+
+        double avg = s.avg();
+        double denominator = usePopulationVariance ? len : len - 1;
+
+        double acc = 0;
+        for (int i = 0; i < len; i++) {
+            final double x = s.getDouble(i);
+            acc += (x - avg) * (x - avg);
+        }
+
+        return acc / denominator;
+    }
+
     public static double ofDoubles(Series<? extends Number> s, boolean usePopulationVariance) {
         return SeriesCompactor.toDoubleSeries(s).variance(usePopulationVariance);
     }
@@ -19,13 +105,19 @@ public class Variance {
 
         Series<BigInteger> noNulls = SeriesCompactor.noNullsSeries(s);
 
-        int size = noNulls.size();
+        int len = noNulls.size();
+        if (len == 0) {
+            return null;
+        } else if (len == 1) {
+            return usePopulationVariance ? BigDecimal.ZERO : null;
+        }
+
         BigDecimal avg = Average.ofBigints_NullsNotExpected(noNulls);
         MathContext mc = DecimalOps.op1Context(avg);
-        BigDecimal denominator = new BigDecimal(usePopulationVariance ? size : size - 1);
+        BigDecimal denominator = new BigDecimal(usePopulationVariance ? len : len - 1);
 
         BigDecimal acc = BigDecimal.ZERO;
-        for (int i = 0; i < size; i++) {
+        for (int i = 0; i < len; i++) {
 
             // TODO: I suspect "avg(..)" above already does BigInteger -> BigDecimal conversion, so maybe worth
             //  implementing this whole thing as "castToDecimal().variance()" ?
@@ -42,16 +134,25 @@ public class Variance {
 
     public static BigDecimal ofDecimals(Series<BigDecimal> s, boolean usePopulationVariance) {
 
-        Series<BigDecimal> noNulls = SeriesCompactor.noNullsSeries(s);
-        int size = noNulls.size();
+        int len = s.size();
+        if (len == 0) {
+            return null;
+        } else if (len == 1) {
+            return usePopulationVariance ? BigDecimal.ZERO : null;
+        }
 
-        BigDecimal avg = Average.ofDecimals_NullsNotExpected(noNulls);
+        BigDecimal avg = Average.ofDecimals(s);
         MathContext mc = DecimalOps.op1Context(avg);
-        BigDecimal denominator = new BigDecimal(usePopulationVariance ? size : size - 1);
 
         BigDecimal acc = BigDecimal.ZERO;
-        for (int i = 0; i < size; i++) {
+        int nonNullSize = len;
+
+        for (int i = 0; i < len; i++) {
             BigDecimal x = s.get(i);
+            if (x == null) {
+                nonNullSize--;
+                continue;
+            }
 
             BigDecimal dev = x.subtract(avg, mc);
             BigDecimal square = dev.multiply(dev, mc);
@@ -59,6 +160,7 @@ public class Variance {
             acc = acc.add(square, mc);
         }
 
+        BigDecimal denominator = new BigDecimal(usePopulationVariance ? nonNullSize : nonNullSize - 1);
         return acc.divide(denominator, mc);
     }
 }